links seem to be working. Streamlining solution.
If the file is in the “fun-stuff” folder, and the linking index is NOT
in that folder, it needs a “/fun-stuff/file” path. If the linking index
IS in that folder, you would think it could just get “/file” but that
didn’t seem to work. I think the “/“ knows it is an absolute root, not
just relative to that folder.